machinst: make get_output_reg target independent;
